Revert "Revert "Revert "Expression test parity safe"""
