+++ src/main/java/com/hubspot/jinjava/lib/exptest/ExpTestLibrary.java
@@ -15,20 +15,6 @@ public class ExpTestLibrary extends SimpleLibrary<ExpTest> {
       IsDefinedExpTest.class,
       IsDivisibleByExpTest.class,
       IsEqualToExpTest.class,
-      IsEqExpTest.class,
-      IsEqualsSymbolExpTest.class,
-      IsNeExpTest.class,
-      IsNotEqualToSymbolExpTest.class,
-      IsLtTest.class,
-      IsLessThanExpTest.class,
-      IsLessThanSymbolExpTest.class,
-      IsLeTest.class,
-      IsLessThanOrEqualToSymbolExpTest.class,
-      IsGtTest.class,
-      IsGreaterThanExpTest.class,
-      IsGreaterThanSymbolExpTest.class,
-      IsGeTest.class,
-      IsGreaterThanOrEqualToSymbolExpTest.class,
       IsEvenExpTest.class,
       IsIterableExpTest.class,
       IsLowerExpTest.class,
@@ -38,20 +24,14 @@ public class ExpTestLibrary extends SimpleLibrary<ExpTest> {
       IsOddExpTest.class,
       IsSameAsExpTest.class,
       IsSequenceExpTest.class,
-      IsBooleanExpTest.class,
-      IsIntegerExpTest.class,
-      IsFloatExpTest.class,
       IsStringExpTest.class,
       IsStringContainingExpTest.class,
       IsStringStartingWithExpTest.class,
-      IsTrueExpTest.class,
-      IsFalseExpTest.class,
       IsTruthyExpTest.class,
       IsUndefinedExpTest.class,
       IsUpperExpTest.class,
       IsContainingAllExpTest.class,
       IsContainingExpTest.class,
-      IsInExpTest.class,
       IsWithinExpTest.class
     );
   }
