--- src/main/java/com/hubspot/jinjava/lib/exptest/IsFloatExpTest.java
@@ -1,35 +1,0 @@
-package com.hubspot.jinjava.lib.exptest;
-
-import com.hubspot.jinjava.doc.annotations.JinjavaDoc;
-import com.hubspot.jinjava.doc.annotations.JinjavaParam;
-import com.hubspot.jinjava.doc.annotations.JinjavaSnippet;
-import com.hubspot.jinjava.interpret.JinjavaInterpreter;
-import java.math.BigDecimal;
-
-@JinjavaDoc(
-  value = "Return true if object is a float",
-  input = @JinjavaParam(value = "value", type = "object", required = true),
-  snippets = {
-    @JinjavaSnippet(
-      code = "{% if num is float %}\n" +
-      "      <!--code to render if num contains an floating point value-->\n" +
-      "{% endif %}"
-    )
-  }
-)
-public class IsFloatExpTest implements ExpTest {
-
-  @Override
-  public String getName() {
-    return "float";
-  }
-
-  @Override
-  public boolean evaluate(Object var, JinjavaInterpreter interpreter, Object... args) {
-    return (
-      var instanceof Double ||
-      var instanceof Float ||
-      (var instanceof BigDecimal && ((BigDecimal) var).scale() > 0)
-    );
-  }
-}
