--- src/main/java/com/hubspot/jinjava/lib/exptest/IsGreaterThanOrEqualToSymbolExpTest.java
@@ -1,12 +1,0 @@
-package com.hubspot.jinjava.lib.exptest;
-
-import com.hubspot.jinjava.doc.annotations.JinjavaDoc;
-
-@JinjavaDoc(value = "", aliasOf = "ge")
-public class IsGreaterThanOrEqualToSymbolExpTest extends IsGeTest {
-
-  @Override
-  public String getName() {
-    return ">=";
-  }
-}
