--- src/main/java/com/hubspot/jinjava/lib/exptest/IsLessThanExpTest.java
@@ -1,12 +1,0 @@
-package com.hubspot.jinjava.lib.exptest;
-
-import com.hubspot.jinjava.doc.annotations.JinjavaDoc;
-
-@JinjavaDoc(value = "", aliasOf = "lt")
-public class IsLessThanExpTest extends IsLtTest {
-
-  @Override
-  public String getName() {
-    return "lessthan";
-  }
-}
