+++ src/main/java/com/hubspot/jinjava/lib/exptest/IsSequenceExpTest.java
@@ -1,7 +1,5 @@
 package com.hubspot.jinjava.lib.exptest;
 
-import static com.hubspot.jinjava.lib.exptest.IsIterableExpTest.isIterable;
-
 import com.hubspot.jinjava.doc.annotations.JinjavaDoc;
 import com.hubspot.jinjava.doc.annotations.JinjavaParam;
 import com.hubspot.jinjava.doc.annotations.JinjavaSnippet;
@@ -27,6 +25,9 @@ public class IsSequenceExpTest implements ExpTest {
 
   @Override
   public boolean evaluate(Object var, JinjavaInterpreter interpreter, Object... args) {
-    return isIterable(var);
+    return (
+      var != null &&
+      (var.getClass().isArray() || Iterable.class.isAssignableFrom(var.getClass()))
+    );
   }
 }
