+++ src/test/java/com/hubspot/jinjava/lib/exptest/IsWithinExpTestTest.java
@@ -26,17 +26,16 @@ public class IsWithinExpTestTest {
       .isEqualTo("pass");
   }
 
-  //    TODO: Uncomment out when ColectionMemberShipOperator.java changes get approved
-  //  @Test
-  //  public void itPassesOnNullValueInSequence() {
-  //    assertThat(
-  //        jinjava.render(
-  //          String.format(IN_TEMPLATE, "null", "[1, 2, null]"),
-  //          new HashMap<>()
-  //        )
-  //      )
-  //      .isEqualTo("pass");
-  //  }
+  @Test
+  public void itFailsOnNullValueInSequence() {
+    assertThat(
+        jinjava.render(
+          String.format(IN_TEMPLATE, "null", "[1, 2, null]"),
+          new HashMap<>()
+        )
+      )
+      .isEqualTo("fail");
+  }
 
   @Test
   public void itFailsOnValueNotInSequence() {
@@ -51,6 +50,12 @@ public class IsWithinExpTestTest {
     assertThat(
         jinjava.render(String.format(IN_TEMPLATE, "null", "[1, 2, 3]"), new HashMap<>())
       )
+      .isEqualTo("fail");
+  }
+
+  @Test
+  public void itFailsOnNullSequence() {
+    assertThat(jinjava.render(String.format(IN_TEMPLATE, "2", "null"), new HashMap<>()))
       .isEqualTo("fail");
   }
